qemu: Store all iothread's 'poll*' attributes as unsigned long long
Convert the internal types to unsigned long long. Luckily we can also
covert the external types too:

 - 'qemuDomainSetIOThreadParams' can accept both _UINT and _ULLONG by
   converting to 'virTypedParamsGetUnsigned'

 - querying is handled via the bulk stats API which is flexible:
    - we use virTypedParamListAddUnsigned to use the bigger type only if
      necessary
    - most users don't even notice because the bindings abstract the
      data types

Apart from the code modifications we also improve the documentation
which was missing for the setters.

+++ src/qemu/qemu_monitor.h
@@ -1254,8 +1254,8 @@ struct _qemuMonitorIOThreadInfo {
     int thread_id;
     bool poll_valid;
     unsigned long long poll_max_ns;
-    unsigned int poll_grow;
-    unsigned int poll_shrink;
+    unsigned long long poll_grow;
+    unsigned long long poll_shrink;
     int thread_pool_min;
     int thread_pool_max;
     bool set_poll_max_ns;
